Held Down Firing
Can now hold down the mouse button to fire and change fire rate in the shooting script on the player. Also made it so bullets have a life span so that they don't hoard memory. May be better to change to an out of bounds box with a box collider but this is a quick fix.